% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-pipeline.R
\name{compile_pipeline}
\alias{compile_pipeline}
\title{Compile a lazy Mongo query into an aggregation pipeline}
\usage{
compile_pipeline(x)
}
\arguments{
\item{x}{A \code{tbl_mongo} object.}
}
\value{
A list of MongoDB aggregation stages.
}
\description{
Compile a lazy Mongo query into an aggregation pipeline
}
\examples{
tbl <- tbl_mongo(
  list(name = "orders"),
  schema = c("status", "amount"),
  executor = function(pipeline, ...) tibble::tibble()
)

query <- dplyr::filter(tbl, amount > 0)
compile_pipeline(query)
}
