% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mdbplyr-package}
\alias{mdbplyr}
\alias{mdbplyr-package}
\title{mdbplyr package}
\description{
A native tidy lazy backend for MongoDB aggregation pipelines.
}
\author{
\strong{Maintainer}: Paolo Bosetti \email{paolo.bosetti@unitn.it}

}
\keyword{internal}
