% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs-arrange.R
\name{arrange.tbl_mongo}
\alias{arrange.tbl_mongo}
\title{Arrange a lazy Mongo query}
\usage{
arrange.tbl_mongo(.data, ..., .by_group = FALSE)
}
\arguments{
\item{.data}{A \code{tbl_mongo} object.}

\item{...}{Bare field names or \code{desc(field)}.}

\item{.by_group}{Whether to prefix the ordering with grouping fields.}
}
\value{
A modified \code{tbl_mongo} object.
}
\description{
Arrange a lazy Mongo query
}
\examples{
tbl <- tbl_mongo(
  list(name = "orders"),
  schema = c("amount"),
  executor = function(pipeline, ...) tibble::tibble()
)

dplyr::arrange(tbl, dplyr::desc(amount))
}
