% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs-group-summarise.R
\name{group_by.tbl_mongo}
\alias{group_by.tbl_mongo}
\title{Group a lazy Mongo query}
\usage{
\method{group_by}{tbl_mongo}(.data, ..., .add = FALSE, .drop = dplyr::group_by_drop_default(.data))
}
\arguments{
\item{.data}{A \code{tbl_mongo} object.}

\item{...}{Bare field names.}

\item{.add}{Whether to add to existing groups.}

\item{.drop}{Included for dplyr compatibility.}
}
\value{
A modified \code{tbl_mongo} object.
}
\description{
Group a lazy Mongo query
}
\examples{
tbl <- tbl_mongo(
  list(name = "orders"),
  schema = c("status", "amount"),
  executor = function(pipeline, ...) tibble::tibble()
)

dplyr::group_by(tbl, status)
}
