% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs-select.R
\name{rename.tbl_mongo}
\alias{rename.tbl_mongo}
\title{Rename fields in a lazy Mongo query}
\usage{
\method{rename}{tbl_mongo}(.data, ...)
}
\arguments{
\item{.data}{A \code{tbl_mongo} object.}

\item{...}{Named bare field renames.}
}
\value{
A modified \code{tbl_mongo} object.
}
\description{
Rename fields in a lazy Mongo query
}
\examples{
tbl <- tbl_mongo(
  list(name = "orders"),
  schema = c("status", "amount"),
  executor = function(pipeline, ...) tibble::tibble()
)

dplyr::rename(tbl, total = amount)
}
