% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs-arrange.R
\name{slice_head.tbl_mongo}
\alias{slice_head.tbl_mongo}
\alias{slice_tail.tbl_mongo}
\alias{head.tbl_mongo}
\title{Slice a lazy Mongo query}
\usage{
slice_head.tbl_mongo(.data, ..., n = NULL, prop = NULL, by = NULL)

slice_tail.tbl_mongo(.data, ..., n = NULL, prop = NULL, by = NULL)

\method{head}{tbl_mongo}(x, n = 6L, ...)
}
\arguments{
\item{.data}{A \code{tbl_mongo} object.}

\item{...}{Must be empty.}

\item{n}{Number of rows to keep. Negative values drop rows from the opposite
end, matching \code{dplyr}.}

\item{prop}{Unsupported.}

\item{by}{Unsupported.}

\item{x}{A \code{tbl_mongo} object.}
}
\value{
A modified \code{tbl_mongo} object.
}
\description{
Slice a lazy Mongo query
}
\examples{
tbl <- tbl_mongo(
  list(name = "orders"),
  schema = c("amount"),
  executor = function(pipeline, ...) tibble::tibble()
)

dplyr::slice_head(tbl, n = 2)
dplyr::slice_tail(tbl, n = 2)
head(tbl, 2)
}
