% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs-mutate.R
\name{transmute.tbl_mongo}
\alias{transmute.tbl_mongo}
\title{Compute and keep only derived fields}
\usage{
\method{transmute}{tbl_mongo}(.data, ...)
}
\arguments{
\item{.data}{A \code{tbl_mongo} object.}

\item{...}{Named scalar expressions.}
}
\value{
A modified \code{tbl_mongo} object.
}
\description{
Compute and keep only derived fields
}
\examples{
tbl <- tbl_mongo(
  list(name = "orders"),
  schema = c("amount"),
  executor = function(pipeline, ...) tibble::tibble()
)

dplyr::transmute(tbl, doubled = amount * 2)
}
