% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_fields}
\alias{schema_fields}
\title{Inspect known fields for a lazy Mongo query}
\usage{
schema_fields(x)
}
\arguments{
\item{x}{A \code{tbl_mongo} or \code{mongo_src} object.}
}
\value{
A character vector of known field names.
}
\description{
Inspect known fields for a lazy Mongo query
}
\examples{
src <- mongo_src(
  list(name = "orders", aggregate = function(...) tibble::tibble()),
  schema = c("status", "amount")
)
tbl <- tbl_mongo(src)

schema_fields(src)
schema_fields(tbl)
}
